Initial commit: just the "annotate" program.
The "annotate" program is sort of a text-based IDE for analyzing
Homeworlds positions. It is primarily an interface to the AI, so
that you can play a game against the computer by running
"annotate Human Computer" and then entering "ai_move" each time
it is the Computer player's move. However, you can also load
arbitrary positions into the program with "annotate -auto", and
then use the "ai_move" and "deepmove" commands to find the best
move, or to find out whether the game is truly lost.