Refactor ai_static_evaluation() to take the preceding move into account.
Rather than simply passing the "after" state to ai_static_evaluation(),
we now pass both the "before" state and the move itself. This doesn't
pessimize anything, because the callers of ai_static_evaluation() were
all doing ApplyMove::or_die() themselves anyway.

This opens up the possibility of hacking ai_static_evaluation() to
return a "Very Unattractive" value for silly moves such as sacrifices
where none of the actions are used, captures where the enemy can
recapture right away, and color-changes on the first turn of the
game. These don't seem to be big issues for the current hand-coded
AI, but they might be big issues for an AI based on a different
paradigm, such as a neural net.