Add several new tests related to pathological catastrophes.
It is worth mentioning that many "pathological catastrophe" tests are
mishandled by the current "winning moves only" logic in findAllMoves().
That logic assumes various things that are not true when the original
state contains overpopulations.  The following tests are mishandled:

 - p0-win-by-avoid-postcatastrophe-3.txt

 - p0-win-by-avoid-postcatastrophe-4.txt

 - p0-win-by-precatastrophe-sacrifice-and-move.txt

 - p0-win-by-precatastrophe.txt

Tomorrow I will start ripping out the entire concept of "catastrophe",
from the ground up. Since in a real game catastrophes are always taken,
it is a needless complication to have them show up as a special type
of SingleAction. The simpler approach is to implicitly catastrophe all
overpopulations at the end of each turn. In the simpler approach, none
of these pathological cases can ever arise, and the "winning moves only"
logic will be correct in its assumptions. I expect that findAllMoves()
will also be significantly faster.