Remove all SDG transcripts involving "cat and mouse" overpopulations.
In these games, one player creates an overpopulation that would
normally be a winning move, but then forgets(?) to issue the
catastrophe order, allowing the game to continue.  Sometimes the
players' comments indicate that they're trying to set up an
experiment to see what the SDG engine will do.  Clearly such
games are no good for training purposes.