Add the method GameState::hasLost(who). Add a warning for overpopulations.
The "annotate" program now gives a warning (in verbose mode only)
whenever a move leaves four or more pieces of a single color in the
same system. The first time this warning is given, it suggests undoing
the last move and making it again with the appropriate "catastrophe"
actions. Subsequent warnings are terser.