Modify SDG transcripts where overpopulations are immediately catastrophed by the other player.
These transcripts contain positions where one player sets up an overpopulation
but does not catastrophe it. The other player immediately causes the catastrophe
at the start of his turn. Therefore the effect is the same as if the first player
had done the catastrophe himself. It will be better for learning purposes if the
transcript attributes the catastrophe action to the first player.