Get findWinningMove() closer to working when overpopulations are involved.
"AllMoves.cc" makes several assumptions that fail to hold when the starting
position contains overpopulations. For example, if you're asked to find a
winning move, it's not sufficient to focus on the enemy's homeworld; you
also need to repair any overpopulations threatening your own homeworld!
So, we now track whether the starting position "contained_overpopulations",
and if it did, then we don't make certain assumptions anymore.

Add the command "win_move" to the "annotate" program. This command invokes
findWinningMove(), which exercises the "win_only" paths in "AllMoves.cc".
Modify "run-tests.sh" to test the "win_move" command as well as "ai_move";
since the commands exercise different paths, it's worth testing both.
Notice that all the tests pass after this commit, but if you revert just
the "AllMoves.cc" changes, four tests fail with the old "win_only" code.

There is still a problem with combine_postcatastrophes(); when not
pruning worse moves, it fails to record moves that leave the attacker's
homeworld fatally overpopulated. This is clearly not a big deal for AI
purposes, but it means that wxgui incorrectly fails to infer the correct
move in some pathological cases.