Add routine reassignNamesToMove().
This routine takes a WholeMove which is legal relative to one
GameState, and replaces all its "where" and "whither" strings so
that it will be legal relative to a second GameState. The two
provided GameStates must be isomorphic (i.e., toComparableString()
must return the same string for each).

This routine will be useful when dealing with transposition tables.