Get wxgui working on Mac OS X.
Mac OS X's command-line "gcc" doesn't support the -march=native option!
How can the "native" architecture not exist? But anyway, we don't really
need a -march= option.

In order to foreground an application on OS X, you have to explicitly
signal that the application is "foregroundable". Otherwise, even when
the window is at the front, the menu bar at the top of your screen will
continue to reflect whatever application is "really" foregrounded, and
the "really" foregrounded application will be the one capturing keypress
events (even though wxgui will be capturing mouse events). So, if we're
on a Mac (as indicated by __WXMAC__), use a magic incantation from
http://wiki.wxwidgets.org/WxMac_Issues
at the top of OnInit() to signal that wxgui is foregroundable.

Fixes issue #1.