Add C++11 support to the Makefile; refactor some verbose code.
The "const" versions of GameState::systemNamed() and GameState::homeworldOf()
are just copies of the "non-const" versions. Make that more obvious (and
easier to refactor in the future if necessary).

Again, full C++11 support (move constructors, emplace vs. insert)
will come later, once I figure out how to deal with WxWidgets on OSX.
Right now the bottleneck appears to be append_move() due to all its
copying of AllMapT::value_type, which is a pair of a std::string
(expensive to copy) and a WholeMove (essentially a std::vector, even
more expensive to copy).