Further improvements to training workflow.
The incantation is now:

    cd superdupergames-archive && ./cook-everything.sh
    # about 13 seconds
    make train-model
    ./train-model encoder-training-set superdupergames-archive/*.cooked
    # about 1 minute
    ./train-model encoder-weights
    # about 10 minutes
    ./train-model evaluator-training-set superdupergames-archive/*.cooked
    # less than 1 minute
    ./train-model evaluator-weights
    # about 16 minutes
    make annotate
    ./annotate Sam Dave

The evaluator's accuracy is now about 69%, which is good enough to
get the first few moves right, but it has a tendency to move its
large ship away from the homeworld, which is bad. (Human players don't
do this, so there's no data with which to teach the net that it leads
to bad consequences.)

The next step should probably be to implement alpha-beta search with
about three moves of lookahead and see if that improves matters.