Don't throw away the "new state" when computing possible moves.
Rather than compute a `vector<WholeMove>` and then have our caller
loop over it calling `ApplyMove::or_die` on each move, we should
just give the caller direct access to the post-move `state` we've
already computed. The easiest way to do this is by passing in a
callback to `findAllMoves`, so that's what we do.

The benchmarks now run in 37, 42, and 49 seconds.
The first benchmark's time is almost halved because it's
testing `ai_move`, and the new code eliminates half of the
work applying moves along that codepath.