Add a Python script to scrape the SDG archive for *all* its games.
This Python script cannot be run as-is; it requires a login cookie from
SuperDuperGames. You can generate a login cookie by logging into SDG with
your web browser of choice and then examining what cookies are being sent
to the server; the one you want will have a key something like

    MasonX-Request-WithApacheSession-cookie

and a value of some long hex string.