Add "cook.py" to replace "cook-everything.sh".
This script is vastly more useful at error reporting, and vastly more
maintainable. (It's also vastly bigger, but such is life.)
Run

    ./cook.py *.raw

to generate *.cooked and *.cooked-but-invalid.
Here are the invalid game transcripts that are not reasonably well
explained by the output of "cook.py":

Game 710
exposes a bug in SDG's (historical) game logic: sac green at
a y1 star and build a y2. The y1 should have been returned to the
stash as soon as the last ship was sacrificed, so building a y2
shouldn't have been possible.

Game 10634
exposes a bug in SDG's (historical) game logic: catastropheing
a homeworld's star did not end the game.

Game 11214
was played with the explicit goal of stress-testing SDG's own
game logic, and ends with catastropheing all four colors at a
single star --- which would be fine except that the catastrophes
are listed in an implausible order, so that the star is destroyed
early on and then the later catastrophes refer to it by name
when it no longer exists.

Games 30216 30758 32153 33661 33671 33760 33762 34373 34375 34437
have "sac y2; move r1 to Defender; cat red at Defender; move g1 from
X to Defender", where X and Defender were not adjacent
prior to the catastrophe. This cannot be expressed
as a valid move if you restrict catastrophes (as we do) to occur only
at end-of-turn boundaries. Interestingly, the attacker in at least
five of these cases (and the defender in one case) is Felix.

Games 32153 32192 32413 32774 32841 32921 33011 33134 33631 33827 33959 34198 34371
have "sac y2; move r2 to X; cat red at X; move r1 to X",
which cannot be expressed as a valid move if you restrict catastrophes
(as we do) to occur only at end-of-turn boundaries.

Games 33323 33369 33901 34154
have "sac y2; move r2 to X; cat red at X; move b1 to Y (g1)", where
the g1 was not in the stash prior to the catastrophe. This cannot
be expressed as a valid move if you restrict catastrophes (as we do)
to occur only at end-of-turn boundaries.

Games 32864 33306 33553 33565 34165
have "sac g2; build r2 at X; cat red at X; build g1 at Y", where
the g1 was not in the stash prior to the catastrophe. This cannot
be expressed as a valid move if you restrict catastrophes (as we do)
to occur only at end-of-turn boundaries.

Game 33970
has "sac b2; trade y2 r2 at X; cat red at X; trade y1 g1 at Y", where
the g1 was not in the stash prior to the catastrophe. This cannot
be expressed as a valid move if you restrict catastrophes (as we do)
to occur only at end-of-turn boundaries.

Games 9329 10575 19179 19419 20159 30076 32480 33917
leave the attacker's homeworld temporarily vacant, by sacrificing
the last yellow ship at home in order to move another ship in.
It is okay to leave the homeworld temporarily undefended, but not
to leave it completely vacant --- that should return it to the stash
and thus count as suicide.

Game 32876
is a very short abandoned game where the first player
passes twice and the second player doesn't; so that the second player
is building ships before the first player has chosen a homeworld setup.