Refactor SingleAction to use Piece instead of Color+Size.
Introduce `Piece::isMissingPieces()` for clarity; it returns true
if the `Piece` has unknown color and/or unknown size.

Also introduce `Piece::empty()` for symmetry with `string::empty()`;
it returns true only if the `Piece` has unknown color AND unknown size,
which is to say, the user didn't tell us anything about the piece.
(`empty()` may or may not be a good name for this accessor, really.)

Simplify `SingleAction::sanitycheck()`.