Add Google Test (GTest) unit tests.
Instructions for installing GTest on Mac OSX are
available here: https://stackoverflow.com/a/46611467/1424877

My intent is that you shouldn't *need* to install GTest simply in order
to compile `annotate`, any more than you *need* to install WxWidgets
if you don't plan to compile `wxgui`. But it's past time for this
project to get some unit tests.

Rip out `GameState::mirror()` as it is unused... and since it was unused,
it didn't work! A missing `&` had caused it to be an identity function
instead of actually mirroring anything.