git rm tests/, move functionality into GTest, and fix a bug.
The old `LEGAL` command would verify not only that the move could be
applied, but also that it was found via `findAllMoves`. Move that logic
into GTest so that it can be ripped out of `annotate`!

The previous version of "win_by_precatastrophe_sacrifice_and_postcatastrophe"
was not as tricky as I had thought. Since Player1 had a b2 at Player2,
all he had to do was sac b2, and then convert r1 to y1 at Player1.
This would leave him with a ship at home post-red-catastrophe.

The new version gives Player1 a g2 at Player2 so that really the only
way to get a non-red ship at home is to fly one there.

This contradicts (as it was meant to contradict) the wisdom in
`findAllMoves` that says we never fly homeward when we're trying
to find wins. Therefore, fix that bug now.