The Heroku app can now parse transcripts!
I'm now finally able to visit the root page,
see a list of games in which speardane is to move,
click to select a game,
see speardane's proposed move,
click to confirm the submission of speardane's move,
and be returned to the home page!

This takes three long page loads, but it works!
No more need to log in as speardane!
(Except to send and/or read in-game comments, unfortunately.)