Eliminate hard aborts in GameState::scan().
The new behavior of `annotate` is only slightly more user-friendly
than before, but the new behavior in Python is much much better.
Before this commit, Python would allow you to create a bogus
GameState with too many pieces, because it was compiled with
assertions turned off. We don't want that!