Fix comments in AllMoves.cc, static-linkage symbols, and eliminate ALLMOVES_USE_SET.
We can assume that `std::unordered_set` exists.
I also assume (in the absence of any evidence to the contrary) that
`unordered_set` will always be at least as fast as `set` for our purposes.

Anonymous-namespace the types in "AllMoves.cc" in preparation for
maybe adding a "Retrograde.cc" following the same internal design.