Scrape new games from SDG.
Game 35846: Draw5Playall specifies a move that first catastrophes
his opponent's homeworld, and then builds g2, even though the catastrophe
should have returned a g1 to the stash. Either SDG is messing up the
catastrophe logic, or SDG is simply abandoning its execution of the
move as soon as the catastrophe happens, failing to notice that the
move also contains extraneous illegal actions.

Game 36147: Pib leaves his homeworld utterly unoccupied in the middle
of a move, but then re-occupies it before the end of the turn. This is
considered valid by SDG and Pyramid Quartet rules, but in my code it
causes the Pib star system to return to the stash and thus the
final action "move y1 from Outpost to Pib" to be rejected because
"Pib" does not identify an existing system.